% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anu-package.R
\docType{package}
\name{anubrand-package}
\alias{anubrand}
\alias{anubrand-package}
\title{anubrand: Tools and templates with Australian National University brand}
\description{
Tools and templates for use at Australian National University.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anuopensci/anubrand}
  \item \url{https://anuopensci.github.io/anubrand/}
  \item Report bugs at \url{https://github.com/anuopensci/anubrand/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
