% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo_get}
\alias{logo_get}
\title{Get the ANU logo}
\source{
https://anu365.sharepoint.com/sites/ANUidentityHUB/SitePages/ANU-logo.aspx
}
\usage{
logo_get(
  path = ".",
  overwrite = FALSE,
  color_mode = c("RGB", "CMYK"),
  category = c("primary", "secondary", "crest", "favicon"),
  type = c("vertical", "horizontal", "inversed", "outline"),
  color = c("gold_black", "gold_white", "black", "white", "gold"),
  ext = c("png", "jpg", "eps"),
  filename = NULL
)
}
\arguments{
\item{path}{The path to save the logo to.}

\item{overwrite}{Overwrite the logo if it already exists.}

\item{color_mode}{Use "RGB" for digital communication (e.g. website) and "CMYK" for printing.}

\item{category}{Either "primary", "secondary", "crest", or "favicon".}

\item{type}{If category is primary or secondary, then "vertical" or "horizontal", otherwise "inversed" or "outline".}

\item{color}{The color of the logo ("gold_black", "gold_white", "black", "white", or "gold"). Note that some colors do not exist based on category and type chosen.}

\item{ext}{The file extension of the logo. Note that some extensions do not exist based on category, type and color chosen.}

\item{filename}{The output filename of the logo (defaults to the original filename).}
}
\description{
A quick way of getting the logo. Not all combinations of category, type, and color (see table below).\tabular{rcccc}{
    \tab primary \tab secondary \tab crest \tab favicon \cr
   vertical \tab Y \tab Y \tab  \tab  \cr
   horizontal \tab Y \tab Y \tab  \tab  \cr
   inversed \tab  \tab  \tab Y \tab Y \cr
   outline \tab  \tab  \tab Y \tab  \cr
   gold_black \tab Y \tab Y \tab  \tab  \cr
   gold_white \tab Y \tab Y \tab  \tab  \cr
   black \tab Y \tab Y \tab Y \tab Y \cr
   white \tab Y \tab Y \tab Y \tab Y \cr
   gold \tab  \tab  \tab Y \tab Y \cr
}
}
